make minimal necessary changes to get NES core building (forked/merged from jgnes)
