+++ Cargo.toml
@@ -36,6 +36,7 @@ serde_json = "1"
 test-log = "0.2"
 thiserror = "1"
 time = "0.3"
+tinyvec = "1"
 toml = "0.8"
 windows = "0.52"
 wgpu = "0.18"
