+++ backend/nes-core/src/api.rs
