+++ backend/nes-core/src/apu/noise.rs
