+++ backend/nes-core/src/apu/pulse.rs
