+++ backend/nes-core/src/bus/cartridge/mappers/konami/irq.rs
