+++ backend/nes-core/src/num.rs
