+++ backend/nes-core/src/serialize.rs
