--- jgnes-core/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "jgnes-core"
-version = "0.6.2"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-jgnes-proc-macros = { path = "../jgnes-proc-macros" }
-
-bincode = { version = "2.0.0-rc.3", features = ["serde"] }
-log = "0.4"
-rand = "0.8"
-serde = { workspace = true }
-thiserror = "1"
-tinyvec = { version = "1", features = ["serde"] }
-
-[lints]
-workspace = true
