--- jgnes-proc-macros/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "jgnes-proc-macros"
-version = "0.6.2"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-proc-macro = true
-
-[dependencies]
-chrono = { version = "0.4", default-features = false, features = ["clock", "std"] }
-quote = "1"
-syn = "2"
-
-[lints]
-workspace = true
